Added network messenger trait. Implemented GrpcMessenger. Moved tests to the separated file
