+++ Cargo.toml
@@ -6,6 +6,8 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+async-trait = "0.1.75"
+bytes = "1.5.0"
 chrono = "0.4.31"
 dslab-mp = { git = "https://github.com/osukhoroslov/dslab", version = "0.1.0" }
 dyn-clone = "1.0.16"
